spring-boot: update user service

+++ backend/spring-boot/user-service/src/main/java/com/my_profile/user_service/service/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.my_profile.user_service.service.impl;
 
 import com.my_profile.user_service.controller.User;
+import com.my_profile.user_service.exception.AccessDbException;
 import com.my_profile.user_service.repository.UserRepository;
 import com.my_profile.user_service.service.UserService;
 import org.springframework.stereotype.Service;
@@ -32,17 +33,16 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User addUser(User user) {
+    public User addUser(User user) throws AccessDbException {
         try{
             return userRepository.insert(user);
         } catch (Exception e){
-            e.printStackTrace();
+            throw new AccessDbException("Add user failed!");
         }
-        return null;
     }
 
     @Override
-    public User updateUser(String id, User user) {
+    public User updateUser(String id, User user) throws AccessDbException {
         Optional<User> optionalUser = this.userRepository.findById(id);
         if(optionalUser.isPresent()){
             User presentUser = optionalUser.get();
@@ -53,17 +53,25 @@ public class UserServiceImpl implements UserService {
             presentUser.setBod(user.getBod());
             presentUser.setMajor(user.getMajor());
             presentUser.setDescription(user.getDescription());
-            return this.userRepository.save(presentUser);
+            try {
+                return this.userRepository.save(presentUser);
+            } catch (Exception e){
+                throw new AccessDbException("Update user failed!");
+            }
         }
         return null;
     }
 
     @Override
-    public User deleteUser(User user) {
-        try{
-            userRepository.delete(user);
-        } catch (Exception e){
-            e.printStackTrace();
+    public User deleteUser(String id) throws AccessDbException {
+        User user = this.userRepository.findByUserID(id).orElse(null);
+        if(user != null){
+            try{
+                userRepository.delete(user);
+                return user;
+            } catch (Exception e){
+                throw new AccessDbException("Delete user failed!");
+            }
         }
         return null;
     }
